requirements.txt
flask
python-dotenv

pinecone-client

langchain
langchain-core
langchain-community
langchain-pinecone
langchain-google-genai

pypdf